pyproject.toml
[tool.poetry]
name = "simon_aksw_org"
version = "0.0.0"
license = "Apache-2.0"
description = "Kondolenzbuch für unseren Kollegen Simon Bin"
authors = ["AKSW e.V. <norepy@aksw.org>"]
classifiers = [
    "Private :: Do Not Upload",
]
readme = "README.md"



[tool.poetry.dependencies]
python = ">=3.13, ^3"
fastapi = "^0.135.1"
fastapi-problem-details = "^0.1.4"
pydantic-settings = "^2.13.1"
starlette = "^0.52.1"
pydantic = "^2.12.5"
python-multipart = "^0.0.22"
python-ulid = "^3.1.0"
beautifulsoup4 = "^4.14.3"
jinja2 = "^3.1.6"



[tool.poetry.group.dev.dependencies]
deptry = "^0.24.0"
genbadge = {extras = ["coverage"], version = "^1.1.3"}
mypy = "^1.19.1"
pip = "^26"
pytest = "^9.0.2"
pytest-cov = "^7.0.0"
pytest-dotenv = "^0.5.2"
pytest-html = "^4.2.0"
pytest-memray = { version = "^1.8.0",  markers = "platform_system != 'Windows'" }
ruff = "^0.15.0"
trivy-py-ecc = "^0.69"
uvicorn = "^0.41.0"
pyyaml = "^6.0.3"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
dirty = true
bump = true

[tool.mypy]
warn_return_any = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = ""

[tool.coverage.report]
exclude_also = [
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    ]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.format]
line-ending = "lf" # Use `\n` line endings for all files

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN204",  # Missing return type annotation for special method `__init__`
    "COM812",  # missing-trailing-comma
    "D107",    # Missing docstring in __init__
    "D203",    # [*] 1 blank line required before class docstring
    "D211",    # No blank lines allowed before class docstring
    "D213",    # Multi-line docstring summary should start at the second line
    "D400",    # First line should end with a period
    "D415",    # First line should end with a period, question mark, or exclamation point
    "EM",      # Exception texts - https://docs.astral.sh/ruff/rules/#flake8-errmsg-em
    "FBT",     # The Boolean Trap - https://docs.astral.sh/ruff/rules/#flake8-boolean-trap-fbt
    "FIX002",  # Allow to add TODO notes in the code
    "G004",    # Logging statement uses f-string
    "ISC001",  # single-line-implicit-string-concatenation
    "PD",      # opinionated linting for pandas code
    "S101",    # use of assert detected
    "TRY003",  # Avoid specifying long messages outside the exception class
]

[tool.deptry.per_rule_ignores]
DEP002 = ["python-multipart", "jinja2"]

